requirements.txt
flask
flask_cors
rembg
pillow